rbp: Update the GL libs to new naming scheme
As the opensource mesa GL library is getting more usable, the name collision wih the firmware GL driver is causing issues.
As such we are renaming the firmware GL driver to avoid this.

The new names are libbrcmEGL.so and libbrcmGLESv2.so. Both will be supported for some time, but the original name
will be dropped at some point